+++ src/lib/components/atom/index.ts
@@ -1,4 +1,5 @@
 export { default as Button } from './Button.svelte';
 export { default as FormAction } from './FormAction.svelte';
+export { default as HasBeenTabbed } from './HasBeenTabbed.svelte';
 export { default as Link } from './Link.svelte';
 export { default as Spinner } from './Spinner.svelte';